requirements.txt
streamlit==1.38.0
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-google-genai==1.0.8
pypdf==4.3.1
chromadb==0.5.5
python-dotenv==1.0.1
google-generativeai==0.7.2
langchain-chroma==0.1.2